+++ CHANGELOG.md
@@ -41,9 +41,12 @@ The [history](HISTORY.md) contains the changelogs of older PostBOUND releases.
   of results to gather for all queries in a more flexible and intuitive way (how did *explain=True*  and *analyze=True*
   interact?).
   The old API using *analyze=True*, etc. is now deprecated in favor of these new parameters.
+- Column references now provide a `drop_table_alias()` method to obtain a normalized-ish representation of the column.
+  This should be helpful in situations where references to the same column are not consistent in their table references,
+  e.g., when one was obtained from the schema and the other was obtained from the query.
+- The `PredicateVisitor` can now be started at the query. It will extract the predicates as needed.
 - The `to_json()` and `to_json_dump()` utilities now support dataclasses out-of-the-box.
 - Expose `argmax()` directly in _util_ module
-- The `PredicateVisitor` can now be started at the query. It will extract the predicates as needed.
 
 ## 🏥 Fixes
 - Fixed output format of the benchmarking log if additional entries are appended to an existing log. Essentially, we fix such